+++ packages/gitbook/src/components/AIChat/AIChatMessages.tsx
@@ -1,17 +1,20 @@
 import { tcls } from '@/lib/tailwind';
 import { AIMessageRole } from '@gitbook/api';
 import type React from 'react';
-import type { AIChatState } from '../AI/useAIChat';
+import type { AIChatController, AIChatState } from '../AI/useAIChat';
+import { AIChatFollowupSuggestions } from './AiChatFollowupSuggestions';
 
 export function AIChatMessages(props: {
     chat: AIChatState;
+    chatController: AIChatController;
     lastUserMessageRef?: React.RefObject<HTMLDivElement>;
 }) {
-    const { chat, lastUserMessageRef } = props;
+    const { chat, chatController, lastUserMessageRef } = props;
 
     return (
         <>
             {chat.messages.map((message, index) => {
+                const isLastMessage = index === chat.messages.length - 1;
                 const isLastUserMessage =
                     message.role === AIMessageRole.User &&
                     index === chat.messages.map((m) => m.role).lastIndexOf(AIMessageRole.User);
@@ -23,8 +26,9 @@ export function AIChatMessages(props: {
                             message.content ? 'animate-[fadeIn_500ms_both]' : '',
                             'shrink-0',
                             'last:min-h-[calc(100%-5rem)]',
+                            'flex flex-col gap-6',
                             message.role === AIMessageRole.User
-                                ? 'max-w-[80%] self-end rounded-md bg-tint px-4 py-2'
+                                ? 'max-w-[80%] self-end circular-corners:rounded-2xl rounded-corners:rounded-md bg-tint px-4 py-2'
                                 : ''
                         )}
                         style={{
@@ -34,7 +38,7 @@ export function AIChatMessages(props: {
                     >
                         {message.content ? (
                             message.content
-                        ) : (
+                        ) : chat.loading ? (
                             <div className="flex w-full animate-[fadeIn_500ms_both] flex-wrap gap-2">
                                 {Array.from({ length: 7 }).map((_, index) => (
                                     <div
@@ -47,7 +51,13 @@ export function AIChatMessages(props: {
                                     />
                                 ))}
                             </div>
-                        )}
+                        ) : null}
+                        {isLastMessage ? (
+                            <AIChatFollowupSuggestions
+                                chat={chat}
+                                chatController={chatController}
+                            />
+                        ) : null}
                     </div>
                 );
             })}